Buffer overrun warnings from GCC 7.1 fixed (nw)

+++ src/devices/bus/lpci/mpc105.cpp
@@ -102,7 +102,7 @@ void mpc105_device::update_memory()
 
 				if (m_bank_base > 0)
 				{
-					sprintf(bank_str,"bank%d",bank + m_bank_base);
+					sprintf(bank_str,"bank%d",uint8_t(bank + m_bank_base));
 					membank(bank_str)->set_base(machine().device<ram_device>(RAM_TAG)->pointer());
 				}
 			}
